refactor: refactor:  the functionality of drawing is placed in separate files
